requirements.txt
python-dotenv
Flask
plotly
adafruit-circuitpython-si7021
adafruit-circuitpython-ahtx0
